# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Todo API",
    "version": "1.0.0",
    "description": "A production-ready REST API for managing todos.\n\n## Features\n- Full CRUD on todos\n- Pagination (`skip` / `limit`) on list\n- Full-text search on title (LIKE)\n- Input validation via Zod (structured error arrays)\n- SQLite persistence via sql.js\n\n## Status values\n| Value | Meaning |\n|-------|---------|\n| `pending` | Not started yet (default) |\n| `in-progress` | Currently being worked on |\n| `done` | Completed |",
    "contact": {
      "name": "lootopia-les-3-dev",
      "url": "https://github.com/lootopia-les-3-dev/todo-api-node"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "components": {
    "schemas": {
      "Todo": {
        "type": "object",
        "description": "Un todo stocké en base de données",
        "required": [
          "id",
          "title",
          "status"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "description": "Identifiant unique auto-incrémenté, généré par la BDD",
            "example": 42,
            "readOnly": true
          },
          "title": {
            "type": "string",
            "description": "Titre court de la tâche (1–200 caractères)",
            "minLength": 1,
            "maxLength": 200,
            "example": "Buy groceries"
          },
          "description": {
            "type": "string",
            "description": "Description longue optionnelle (max 1000 caractères). Peut être null.",
            "maxLength": 1000,
            "nullable": true,
            "example": "Oat milk, sourdough bread, eggs"
          },
          "status": {
            "type": "string",
            "description": "Statut actuel du cycle de vie du todo",
            "enum": [
              "pending",
              "in-progress",
              "done"
            ],
            "default": "pending",
            "example": "in-progress"
          }
        },
        "example": {
          "id": 42,
          "title": "Buy groceries",
          "description": "Oat milk, sourdough bread, eggs",
          "status": "in-progress"
        }
      },
      "CreateTodoBody": {
        "type": "object",
        "description": "Payload pour créer un nouveau todo. Seul `title` est obligatoire.",
        "required": [
          "title"
        ],
        "properties": {
          "title": {
            "type": "string",
            "description": "Titre de la tâche (1–200 caractères)",
            "minLength": 1,
            "maxLength": 200,
            "example": "Finish the project report"
          },
          "description": {
            "type": "string",
            "description": "Détails optionnels sur la tâche (max 1000 caractères)",
            "maxLength": 1000,
            "nullable": true,
            "example": "Include Q3 metrics and executive summary"
          },
          "status": {
            "type": "string",
            "description": "Statut initial — `pending` si omis",
            "enum": [
              "pending",
              "in-progress",
              "done"
            ],
            "default": "pending",
            "example": "pending"
          }
        },
        "example": {
          "title": "Finish the project report",
          "description": "Include Q3 metrics and executive summary",
          "status": "pending"
        }
      },
...
      "ValidationError": {
        "type": "object",
        "description": "Retourné quand le body ou les query params échouent la validation Zod.",
        "properties": {
          "detail": {
            "description": "Tableau d'issues Zod — chaque issue a un `path` et un `message`",
            "oneOf": [
              {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "path": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "example": [
                        "title"
                      ]
                    },
                    "message": {
                      "type": "string",
                      "example": "String must contain at least 1 character(s)"
                    }
                  }
                }
              },
              {
                "type": "string"
              }
            ],
            "example": [
              {
                "path": [
                  "title"
                ],
                "message": "Required"
              },
              {
                "path": [
                  "status"
                ],
                "message": "Invalid enum value. Expected 'pending' | 'in-progress' | 'done'"
              }
            ]
          }
        }
      },
...
    },
...
  },
  "paths": {
...
    "/todos": {
      "post": {
        "tags": [
          "Todos"
        ],
        "summary": "Créer un todo",
        "description": "Crée un nouveau todo et retourne l'objet persisté avec son `id` généré automatiquement.\n\n- `title` est le seul champ **obligatoire** (1–200 caractères).\n- `description` vaut `null` par défaut (max 1000 caractères).\n- `status` vaut `pending` par défaut.\n",
        "operationId": "createTodo",
        "requestBody": {
          "required": true,
          "description": "Données du todo. Seul `title` est obligatoire.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTodoBody"
              },
              "examples": {
                "minimal": {
                  "summary": "Titre uniquement (cas minimal)",
                  "value": {
                    "title": "Read a book"
                  }
                },
                "avecDescription": {
                  "summary": "Titre + description",
                  "value": {
                    "title": "Prepare sprint review",
                    "description": "Slides, demo environment, metrics from last sprint"
                  }
                },
                "avecStatut": {
                  "summary": "Tâche déjà en cours",
                  "value": {
                    "title": "Refactor authentication module",
                    "description": "Extract JWT logic into a dedicated service",
                    "status": "in-progress"
                  }
                },
                "dejaDone": {
                  "summary": "Tâche créée comme terminée",
                  "value": {
                    "title": "Write unit tests for database module",
                    "status": "done"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Todo créé avec succès — retourne l'objet complet avec son `id`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                },
                "examples": {
                  "minimal": {
                    "summary": "Créé depuis un payload minimal",
                    "value": {
                      "id": 7,
                      "title": "Read a book",
                      "description": null,
                      "status": "pending"
                    }
                  },
                  "avecDescription": {
                    "summary": "Créé avec description",
                    "value": {
                      "id": 8,
                      "title": "Prepare sprint review",
                      "description": "Slides, demo environment, metrics from last sprint",
                      "status": "pending"
                    }
                  },
                  "enCours": {
                    "summary": "Créé en statut in-progress",
                    "value": {
                      "id": 9,
                      "title": "Refactor authentication module",
                      "description": "Extract JWT logic into a dedicated service",
                      "status": "in-progress"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Échec de validation — title manquant, trop long, ou status invalide",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                },
                "examples": {
                  "titleManquant": {
                    "summary": "title absent du body",
                    "value": {
                      "detail": [
                        {
                          "path": [
                            "title"
                          ],
                          "message": "title is required"
                        }
                      ]
                    }
                  },
                  "titleVide": {
                    "summary": "title vide",
                    "value": {
                      "detail": [
                        {
                          "path": [
                            "title"
                          ],
                          "message": "String must contain at least 1 character(s)"
                        }
                      ]
                    }
                  },
                  "statutInvalide": {
                    "summary": "status non reconnu",
                    "value": {
                      "detail": [
                        {
                          "path": [
                            "status"
                          ],
                          "message": "Invalid enum value. Expected 'pending' | 'in-progress' | 'done', received 'todo'"
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Todos"
        ],
        "summary": "Lister les todos",
        "description": "Retourne un tableau paginé de tous les todos, triés par ordre d'insertion.\n\n**Pagination :**\n- `GET /todos` → page 1 (10 premiers)\n- `GET /todos?skip=0&limit=5` → 5 premiers\n- `GET /todos?skip=5&limit=5` → items 6 à 10\n- `GET /todos?skip=20&limit=100` → au plus 100 items à partir du 21ème\n\nRetourne `[]` si la base est vide ou si `skip` dépasse le nombre total.\n",
        "operationId": "listTodos",
        "parameters": [
          {
            "in": "query",
            "name": "skip",
            "description": "Nombre d'éléments à ignorer (offset). Défaut 0.",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "example": 0
            }
          },
          {
            "in": "query",
            "name": "limit",
            "description": "Nombre maximum d'éléments à retourner (1–100). Défaut 10.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "example": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste paginée des todos (tableau vide si aucun ou si offset dépasse le total)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Todo"
                  }
                },
                "examples": {
                  "premierePage": {
                    "summary": "Première page — todos variés",
                    "value": [
                      {
                        "id": 1,
                        "title": "Buy groceries",
                        "description": "Oat milk, eggs",
                        "status": "pending"
                      },
                      {
                        "id": 2,
                        "title": "Call the dentist",
                        "description": "Tuesday at 10am",
                        "status": "in-progress"
                      },
                      {
                        "id": 3,
                        "title": "Write unit tests",
                        "description": null,
                        "status": "done"
                      }
                    ]
                  },
                  "pageVide": {
                    "summary": "Offset dépasse le total — retourne []",
                    "value": []
                  },
                  "baseVide": {
                    "summary": "Aucun todo en base",
                    "value": []
                  }
                }
              }
            }
          },
          "400": {
            "description": "Paramètres de pagination invalides",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                },
                "examples": {
                  "limitTropEleve": {
                    "summary": "limit supérieur à 100",
                    "value": {
                      "detail": [
                        {
                          "path": [
                            "limit"
                          ],
                          "message": "Number must be less than or equal to 100"
                        }
                      ]
                    }
                  },
                  "limitNegatif": {
                    "summary": "limit inférieur à 1",
                    "value": {
                      "detail": [
                        {
                          "path": [
                            "limit"
                          ],
                          "message": "Number must be greater than or equal to 1"
                        }
                      ]
                    }
                  },
                  "skipNegatif": {
                    "summary": "skip négatif",
                    "value": {
                      "detail": [
                        {
                          "path": [
                            "skip"
                          ],
                          "message": "Number must be greater than or equal to 0"
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}